Update items-by-bib query due to production performance issue
Responding to performance issues observed in production: Update the query used to fetch items for a bib. There's a bug in the ItemService handler that causes items fetched via the `items?bibId={bibId}&nyplSource={nyplSource}` route to take much longer than the theoretically equivalent `bibs/{nyplSource}/{bibId}/items` route. So swapping that out in lieu of investigating the performance bug.